requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
chardet==5.2.0
colorama==0.4.6
configparser==6.0.0
crayons==0.4.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
more-itertools==10.1.0
packaging==23.2
pluggy==1.3.0
pyparsing==3.1.1
pytest==7.4.2
requests==2.31.0
selenium==4.14.0
six==1.16.0
toml==0.10.2
urllib3==2.0.6
wcwidth==0.2.8
webdriver-manager==4.0.1
xlrd==2.0.1
zipp==3.17.0
